+++ app/controllers/transactions_controller.rb
@@ -1,7 +1,8 @@
 class TransactionsController < ApplicationController
 
   def index
-    @transactions = Transaction.all
+    # use Eager loading  to solve N + 1 query performance problem
+    @transactions = Transaction.includes(:manager).all
   end
 
   def show
@@ -10,29 +11,22 @@ class TransactionsController < ApplicationController
 
   def new
     @transaction = Transaction.new
-    @manager = Manager.all.sample
-
-    render "new_#{params[:type]}"
-  end
-
-  def new_large
-    @transaction = Transaction.new
-  end
-
-  def new_extra_large
-    @transaction = Transaction.new
-    @manager = Manager.all.sample
   end
 
   def create
-    @transaction = Transaction.new(params[:transaction].permit!)
-
-    @manager = Manager.all.sample if params[:type] == 'extra'
+    @transaction = Transaction.new(transaction_params)
 
     if @transaction.save
       redirect_to @transaction
     else
-      render "new_#{params[:type]}"
+      render :new
     end
   end
+
+  private
+
+  def transaction_params
+    params[:transaction].permit(:first_name, :last_name, :from_amount, :from_currency, :to_currency)
+  end
+
 end
