+++ app/views/transactions/_form.html.erb
@@ -1,12 +1,7 @@
-<h1>Create 100-1000 USD Transaction</h1>
-<ul class="links-list">
-  <li><%= link_to "Create 0-100 USD Transaction", new_transaction_path(type: 'small') %></li>
-  <li><%= link_to "Create >1000 USD Transaction", new_transaction_path(type: 'extra') %></li>
-</ul>
+<h1>Create Transaction</h1>
 
 <%= form_with model: @transaction do |form| %>
-  <%= form.hidden_field :from_currency, value: 'USD' %>
-  <%= hidden_field_tag :type, 'large' %>
+
   <div class="form-group">
     <%= form.label :first_name %>
     <%= form.text_field :first_name, autocomplete: 'off' %>
@@ -24,7 +19,7 @@
   </div>
 
   <div class="form-group">
-    <%= form.label :from_amount, 'Amount(USD)' %>
+    <%= form.label :from_amount, 'Amount' %>
     <%= form.text_field :from_amount, autocomplete: 'off' %>
     <% @transaction.errors.full_messages_for(:from_amount).each do |message| %>
       <div class="error"><%= message %></div>
@@ -32,8 +27,16 @@
   </div>
 
   <div class="form-group">
-    <%= form.label :to_currency %>
-    <%= form.select :to_currency, Transaction::AVAILABLE_CURRENCIES - ['USD'] %>
+    <%= form.label :from_currency %>
+    <%= form.select :from_currency, Transaction::AVAILABLE_CURRENCIES %>
+    <% @transaction.errors.full_messages_for(:from_currency).each do |message| %>
+      <div class="error"><%= message %></div>
+    <% end %>
+  </div>
+
+  <div class="form-group">
+    <%= form.label :to_currency %><br>
+    <%= form.select :to_currency, Transaction::AVAILABLE_CURRENCIES %>
     <% @transaction.errors.full_messages_for(:to_currency).each do |message| %>
       <div class="error"><%= message %></div>
     <% end %>
@@ -43,4 +46,3 @@
     <%= form.submit %>
   </div>
 <% end %>
-
