+++ app/views/transactions/index.html.erb
@@ -26,9 +26,7 @@
             <%= transaction.client_full_name %>
           </td>
           <td>
-            <% if transaction.manager %>
-              <%= transaction.manager.full_name %>
-            <% end %>
+            <%= transaction.manager&.full_name %>
           </td>
         </tr>
       <% end %>
@@ -36,8 +34,6 @@
   </table>
 
   <ul class="links-list">
-    <li><%= link_to "Create 0-100 USD Transaction", new_transaction_path(type: 'small') %></li>
-    <li><%= link_to "Create 100-1000 USD Transaction", new_transaction_path(type: 'large') %></li>
-    <li><%= link_to "Create >1000 USD Transaction", new_transaction_path(type: 'extra') %></li>
+    <li><%= link_to "Create Transaction", new_transaction_path %></li>
   </ul>
 </div>
