+++ app/views/transactions/new.html.erb
@@ -1,0 +1,1 @@
+<%= render partial: "form" %>
