--- app/views/transactions/new_small.html.erb
@@ -1,30 +1,0 @@
-<h1>Create 0-100 USD Transaction</h1>
-<ul class="links-list">
-  <li><%= link_to "Create 100-1000 USD Transaction", new_transaction_path(type: 'large') %></li>
-  <li><%= link_to "Create >1000 USD Transaction", new_transaction_path(type: 'extra') %></li>
-</ul>
-
-<%= form_with model: @transaction do |form| %>
-  <%= form.hidden_field :from_currency, value: 'USD' %>
-  <%= hidden_field_tag :type, 'small' %>
-  <div class="form-group">
-    <%= form.label :from_amount, 'Amount(USD)' %>
-    <%= form.text_field :from_amount, autocomplete: 'off' %>
-    <% @transaction.errors.full_messages_for(:from_amount).each do |message| %>
-      <div class="error"><%= message %></div>
-    <% end %>
-  </div>
-
-  <div class="form-group">
-    <%= form.label :to_currency %>
-    <%= form.select :to_currency, Transaction::AVAILABLE_CURRENCIES - ['USD'] %>
-    <% @transaction.errors.full_messages_for(:to_currency).each do |message| %>
-      <div class="error"><%= message %></div>
-    <% end %>
-  </div>
-
-  <div>
-    <%= form.submit %>
-  </div>
-<% end %>
-
