+++ config/routes.rb
@@ -1,7 +1,5 @@
 Rails.application.routes.draw do
-  root "transactions#new", type: 'small'
+  root "transactions#index"
 
-  resources :transactions, only: [:index, :show, :create]
-
-  get 'transactions/new/:type', to: 'transactions#new', as: :new_transaction
+  resources :transactions, only: [:index, :show, :create, :new]
 end
